Updated android maven plugin
The older plugin is incompatible with newer releases of gradle.
Newer gradle compatibility is required for DroidDriver to be built
as a subproject of projects that require a newer version of gradle.

com.github.dcendents:android-maven-plugin:1.2 is compatible with
gradle version n where 2.2 <= n < 2.4

com.github.dcendents:android-maven-gradle-plugin:1.3 is compatible
with gradle 2.4+

Change-Id: I42079dc0da92188b0258179fc7ca5ce0dc9a49fc